compliance: add functions for crypitc reading/writing events
