--- internal/service/model_transform.go
@@ -1,35 +1,0 @@
-// Licensed to The Moov Authors under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. The Moov Authors licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package service
-
-type TransformConfig struct {
-	Encoding   *EncodingConfig
-	Encryption *EncryptionConfig
-}
-
-type EncodingConfig struct {
-	Base64 bool
-}
-
-type EncryptionConfig struct {
-	AES *AESConfig
-}
-
-type AESConfig struct {
-	Key string
-}
